iommu: purecap support added

+++ sys/arm64/iommu/smmu.c
@@ -389,7 +389,7 @@ smmu_dump_ste(struct smmu_softc *sc, int sid)
 
 	if (sc->features & SMMU_FEATURE_2_LVL_STREAM_TABLE) {
 		i = sid >> STRTAB_SPLIT;
-		l1 = (void *)((uint64_t)strtab->vaddr +
+		l1 = (void *)((uintptr_t)strtab->vaddr +
 		    STRTAB_L1_DESC_DWORDS * 8 * i);
 		device_printf(sc->dev, "L1 ste == %lx\n", l1[0]);
 
@@ -398,7 +398,7 @@ smmu_dump_ste(struct smmu_softc *sc, int sid)
 		if (ste == NULL) /* L2 is not initialized */
 			return;
 	} else {
-		ste = (void *)((uint64_t)strtab->vaddr +
+		ste = (void *)((uintptr_t)strtab->vaddr +
 		    sid * (STRTAB_STE_DWORDS << 3));
 	}
 
@@ -429,7 +429,7 @@ smmu_evtq_dequeue(struct smmu_softc *sc, uint32_t *evt)
 	evtq = &sc->evtq;
 
 	evtq->lc.val = bus_read_8(sc->res[0], evtq->prod_off);
-	entry_addr = (void *)((uint64_t)evtq->vaddr +
+	entry_addr = (void *)((uintptr_t)evtq->vaddr +
 	    evtq->lc.cons * EVTQ_ENTRY_DWORDS * 8);
 	memcpy(evt, entry_addr, EVTQ_ENTRY_DWORDS * 8);
 	evtq->lc.cons = smmu_q_inc_cons(evtq);
@@ -440,7 +440,7 @@ static void
 smmu_print_event(struct smmu_softc *sc, uint32_t *evt)
 {
 	struct smmu_event *ev;
-	uintptr_t input_addr;
+	uint64_t input_addr;
 	uint8_t event_id;
 	device_t dev;
 	int sid;
@@ -553,7 +553,7 @@ smmu_cmdq_enqueue_cmd(struct smmu_softc *sc, struct smmu_cmdq_entry *entry)
 	} while (smmu_q_has_space(cmdq) == 0);
 
 	/* Write the command to the current prod entry. */
-	entry_addr = (void *)((uint64_t)cmdq->vaddr +
+	entry_addr = (void *)((uintptr_t)cmdq->vaddr +
 	    Q_IDX(cmdq, cmdq->lc.prod) * CMDQ_ENTRY_DWORDS * 8);
 	memcpy(entry_addr, cmd, CMDQ_ENTRY_DWORDS * 8);
 
@@ -594,7 +594,7 @@ smmu_sync(struct smmu_softc *sc)
 	smmu_cmdq_enqueue_cmd(sc, &cmd);
 
 	/* Wait for the sync completion. */
-	base = (void *)((uint64_t)q->vaddr +
+	base = (void *)((uintptr_t)q->vaddr +
 	    Q_IDX(q, prod) * CMDQ_ENTRY_DWORDS * 8);
 
 	/*
@@ -790,7 +790,7 @@ smmu_get_ste_addr(struct smmu_softc *sc, int sid)
 		addr = l1_desc->va;
 		addr += (sid & ((1 << STRTAB_SPLIT) - 1)) * STRTAB_STE_DWORDS;
 	} else {
-		addr = (void *)((uint64_t)strtab->vaddr +
+		addr = (void *)((uintptr_t)strtab->vaddr +
 		    STRTAB_STE_DWORDS * 8 * sid);
 	};
 
@@ -1038,7 +1038,7 @@ smmu_init_l1_entry(struct smmu_softc *sc, int sid)
 	l1_desc->pa = vtophys(l1_desc->va);
 
 	i = sid >> STRTAB_SPLIT;
-	addr = (void *)((uint64_t)strtab->vaddr +
+	addr = (void *)((uintptr_t)strtab->vaddr +
 	    STRTAB_L1_DESC_DWORDS * 8 * i);
 
 	/* Install the L1 entry. */
@@ -1061,7 +1061,7 @@ smmu_deinit_l1_entry(struct smmu_softc *sc, int sid)
 	strtab = &sc->strtab;
 
 	i = sid >> STRTAB_SPLIT;
-	addr = (void *)((uint64_t)strtab->vaddr +
+	addr = (void *)((uintptr_t)strtab->vaddr +
 	    STRTAB_L1_DESC_DWORDS * 8 * i);
 	*addr = 0;
